feat: implement core Gaia CLI architecture and schema definitions for skill tree management
