+++ .github/workflows/publish-pypi.yml
@@ -1,0 +1,39 @@
+name: Publish gaia-cli to PyPI
+
+on:
+  push:
+    tags:
+      - 'gaia-cli-v*'
+  workflow_dispatch: {}
+
+permissions:
+  contents: read
+  id-token: write
+
+jobs:
+  publish:
+    name: Build and Publish
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p/gaia-cli
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.12'
+
+      - name: Install build backend
+        run: python -m pip install build
+
+      - name: Clean local build artifacts
+        run: rm -rf build dist *.egg-info src/*.egg-info
+
+      - name: Build sdist and wheel
+        run: python -m build
+
+      - name: Publish to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
