+++ .gitignore
@@ -32,3 +32,4 @@ build/
 # Temp files
 *.tmp
 *.bak
+scratch/
