+++ plugin/cli/prWriter.py
@@ -1,181 +1,1 @@
-import os
-import re
-import shutil
-import subprocess
-
-
-def open_pr(username, tree_data, candidate_result=None):
-    branch_name = f"update-tree-{username}"
-    if candidate_result:
-        title = f"[user] Update {username} skill tree - added {candidate_result}"
-    else:
-        title = f"[user] Update {username} skill tree"
-    body = f"Detected new skills and confirmed fusions for {username}.\n\nAutomatically generated by Gaia plugin."
-    print("Mock: Opening PR to gaia-registry/gaia")
-    print(f"Branch: {branch_name}")
-    print(f"Title: {title}")
-    print(f"Body: {body}")
-    print("Success: PR #42 opened.")
-    return "https://github.com/gaia-registry/gaia/pull/42"
-
-
-def _sanitize_branch_component(value):
-    safe = re.sub(r"[^a-zA-Z0-9._-]+", "-", (value or "").strip()).strip("-").lower()
-    return safe or "unknown"
-
-
-def _build_similarity_index(similarity_rows):
-    indexed = {}
-    for row in similarity_rows or []:
-        source = row.get("sourceSkillId")
-        target = row.get("targetSkillId")
-        score = row.get("score")
-        if not source or not target:
-            continue
-        indexed.setdefault(source, []).append((score, target))
-    for source in indexed:
-        indexed[source].sort(key=lambda item: item[0], reverse=True)
-    return indexed
-
-
-def build_intake_pr_body(batch_data):
-    source_repo = batch_data.get("sourceRepo", "unknown")
-    generated_at = batch_data.get("generatedAt", "unknown")
-    known_skills = batch_data.get("knownSkills", [])
-    proposed_skills = batch_data.get("proposedSkills", [])
-    similarity = _build_similarity_index(batch_data.get("similarity", []))
-
-    lines = [
-        "## Gaia Draft Skill Intake",
-        "",
-        "This PR submits a draft intake batch for human review.",
-        "The canonical intake artifact is the JSON file in `intake/skill-batches/`.",
-        "Accepted skills must be promoted separately into `graph/gaia.json`.",
-        "",
-        "### Batch Summary",
-        f"- User: `{batch_data.get('userId', 'unknown')}`",
-        f"- Source repo: `{source_repo}`",
-        f"- Generated at: `{generated_at}`",
-        f"- Known skills: `{len(known_skills)}`",
-        f"- Proposed skills: `{len(proposed_skills)}`",
-        "",
-        "### Proposed Skills",
-        "| ID | Name | Type | Similarity hints |",
-        "|---|---|---|---|",
-    ]
-
-    if proposed_skills:
-        for skill in proposed_skills:
-            skill_id = skill.get("id", "")
-            hints = similarity.get(skill_id, [])[:2]
-            if hints:
-                hint_text = ", ".join([f"`{target}` ({score:.3f})" for score, target in hints])
-            else:
-                hint_text = "-"
-            lines.append(
-                f"| `{skill_id}` | {skill.get('name', '')} | `{skill.get('type', '')}` | {hint_text} |"
-            )
-    else:
-        lines.append("| - | - | - | - |")
-
-    lines.extend(
-        [
-            "",
-            "### Reviewer Checklist",
-            "- [ ] Reviewed each proposed skill for clarity and non-overlap.",
-            "- [ ] Marked each candidate as: accept, rename, duplicate, needs evidence, or reject.",
-            "- [ ] Verified similarity hints are directionally useful (not authoritative).",
-            "- [ ] Noted any evidence gaps or requests for follow-up.",
-            "",
-            "### Maintainer Promotion Checklist",
-            "- [ ] Convert accepted draft skills into canonical `graph/gaia.json` edits in a follow-up PR.",
-            "- [ ] Run `python3 scripts/validate.py` and `python3 scripts/validate_intake.py` on promotion PR.",
-            "- [ ] Close or link back this intake PR from the promotion PR.",
-            "",
-            "Automatically generated by `gaia push`.",
-        ]
-    )
-    return "\n".join(lines)
-
-
-def _run(cmd, cwd):
-    return subprocess.run(cmd, cwd=cwd, capture_output=True, text=True, check=False)
-
-
-def _gh_ready(cwd):
-    if shutil.which("gh") is None:
-        return False, "GitHub CLI (`gh`) is not installed."
-    auth = _run(["gh", "auth", "status"], cwd=cwd)
-    if auth.returncode != 0:
-        return False, "GitHub CLI is not authenticated."
-    return True, ""
-
-
-def _print_manual_pr_steps(branch_name, title, body_path, batch_path):
-    print("PR creation skipped. Run these commands manually:")
-    print(f"  git checkout -B {branch_name}")
-    print(f"  git add {batch_path}")
-    print('  git commit -m "intake: add skill batch <batchId>"')
-    print(f"  gh pr create --draft --title \"{title}\" --body-file {body_path} --label draft-skills --label needs-review")
-
-
-def open_intake_pr(username, batch_data, batch_path=None, repo_root="."):
-    if not batch_path:
-        print("No batch path provided; skipping PR creation.")
-        return None
-
-    user_component = _sanitize_branch_component(username)
-    batch_component = _sanitize_branch_component(batch_data.get("batchId"))
-    branch_name = f"intake/{user_component}/{batch_component}"
-    title = f"[intake] {username} skill batch - {batch_data.get('sourceRepo')}"
-    body = build_intake_pr_body(batch_data)
-    body_path = os.path.join(repo_root, ".gaia-intake-pr-body.md")
-    with open(body_path, "w", encoding="utf-8") as f:
-        f.write(body)
-        f.write("\n")
-
-    ready, reason = _gh_ready(repo_root)
-    if not ready:
-        print(reason)
-        _print_manual_pr_steps(branch_name, title, body_path, batch_path)
-        return None
-
-    batch_relpath = os.path.relpath(batch_path, repo_root)
-    _run(["git", "checkout", "-B", branch_name], cwd=repo_root)
-    _run(["git", "add", batch_relpath], cwd=repo_root)
-    commit = _run(
-        ["git", "commit", "-m", f"intake: add skill batch {batch_data.get('batchId')}"],
-        cwd=repo_root,
-    )
-    if commit.returncode != 0 and "nothing to commit" not in commit.stderr.lower():
-        print("Failed to create intake commit.")
-        print(commit.stderr.strip() or commit.stdout.strip())
-        _print_manual_pr_steps(branch_name, title, body_path, batch_path)
-        return None
-
-    pr = _run(
-        [
-            "gh",
-            "pr",
-            "create",
-            "--draft",
-            "--title",
-            title,
-            "--body-file",
-            body_path,
-            "--label",
-            "draft-skills",
-            "--label",
-            "needs-review",
-        ],
-        cwd=repo_root,
-    )
-    if pr.returncode != 0:
-        print("Failed to open PR with gh.")
-        print(pr.stderr.strip() or pr.stdout.strip())
-        _print_manual_pr_steps(branch_name, title, body_path, batch_path)
-        return None
-
-    pr_url = pr.stdout.strip().splitlines()[-1] if pr.stdout.strip() else ""
-    print(f"Success: intake PR opened ({pr_url}).")
-    return pr_url or None
+from gaia_cli.prWriter import *  # noqa: F401,F403
