+++ plugin/cli/resolver.py
@@ -1,19 +1,1 @@
-import json
-import os
-
-def load_canonical_skills(registry_path="graph/gaia.json"):
-    if not os.path.exists(registry_path):
-        pass
-    canonical_skills = set()
-    try:
-        with open(registry_path, 'r') as f:
-            data = json.load(f)
-            for skill in data.get('skills', []):
-                canonical_skills.add(skill['id'])
-    except Exception:
-        pass
-    return canonical_skills
-
-def resolve_skills(detected_tokens, registry_path="graph/gaia.json"):
-    canonical = load_canonical_skills(registry_path)
-    return list(set(detected_tokens).intersection(canonical))
+from gaia_cli.resolver import *  # noqa: F401,F403
