+++ src/gaia_cli/__main__.py
@@ -1,0 +1,5 @@
+from gaia_cli.main import main
+
+
+if __name__ == "__main__":
+    main()
