+++ src/gaia_cli/data/graph/named/devin-ai/autonomous-swe.md
@@ -1,0 +1,28 @@
+---
+id: devin-ai/autonomous-swe
+name: Autonomous SWE
+contributor: devin-ai
+origin: true
+genericSkillRef: autonomous-debug
+status: named
+level: III
+description: Autonomous software engineering agent capable of end-to-end debugging, code generation, and self-correction across complex multi-file codebases.
+links:
+  github: https://github.com/cognition-labs/devin
+tags:
+  - software-engineering
+  - autonomous
+  - debugging
+  - code-generation
+  - self-correction
+createdAt: "2026-04-29"
+updatedAt: "2026-04-29"
+---
+
+## Overview
+
+Autonomous SWE is an autonomous software engineering agent that combines debugging, code generation, and self-correction to resolve issues end-to-end. It navigates complex, multi-file repositories, identifies root causes, proposes and applies fixes, runs tests, and iterates until the problem is resolved—without human intervention.
+
+## Origin
+
+First published by @devin-ai (Cognition Labs). This is the origin implementation for the "autonomous-debug" skill bucket.
