+++ src/gaia_cli/data/graph/named/karpathy/autoresearch.md
@@ -1,0 +1,26 @@
+---
+id: karpathy/autoresearch
+name: AutoResearch
+contributor: karpathy
+origin: true
+genericSkillRef: autonomous-research-agent
+status: named
+level: II
+description: Autonomous research agent that iteratively searches, reads, and synthesizes academic papers into structured summaries.
+links:
+  github: https://github.com/karpathy/autoresearch
+tags:
+  - research
+  - autonomous
+  - paper-synthesis
+createdAt: "2026-04-29"
+updatedAt: "2026-04-29"
+---
+
+## Overview
+
+AutoResearch is an autonomous agent that performs iterative literature review by searching academic databases, reading papers, extracting key findings, and producing structured research summaries.
+
+## Origin
+
+First published by @karpathy. This is the origin implementation for the "autonomous-research-agent" skill bucket.
