+++ src/gaia_cli/data/schema/skillTree.schema.json
@@ -1,0 +1,134 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://gaia-registry.github.io/schema/skillTree.schema.json",
+  "title": "Gaia User Skill Tree",
+  "description": "Schema for a user's personal skill tree record in the Gaia registry.",
+  "type": "object",
+  "required": [
+    "userId",
+    "updatedAt",
+    "unlockedSkills",
+    "pendingCombinations",
+    "stats"
+  ],
+  "additionalProperties": false,
+  "properties": {
+    "userId": {
+      "type": "string",
+      "pattern": "^[a-zA-Z0-9][a-zA-Z0-9-]*$",
+      "description": "GitHub username of the skill tree owner."
+    },
+    "updatedAt": {
+      "type": "string",
+      "format": "date",
+      "description": "ISO 8601 date of last update."
+    },
+    "unlockedSkills": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/unlockedSkill"
+      },
+      "description": "Skills the user has confirmed and unlocked."
+    },
+    "pendingCombinations": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/pendingCombination"
+      },
+      "description": "Detected but unconfirmed combination candidates."
+    },
+    "stats": {
+      "$ref": "#/definitions/treeStats",
+      "description": "Aggregate statistics for the skill tree."
+    }
+  },
+  "definitions": {
+    "unlockedSkill": {
+      "type": "object",
+      "required": ["skillId", "level", "unlockedAt", "unlockedIn"],
+      "additionalProperties": false,
+      "properties": {
+        "skillId": {
+          "type": "string",
+          "pattern": "^[a-z][a-z0-9]*(-[a-z0-9]+)*$",
+          "description": "Reference to a skill ID in the Gaia graph."
+        },
+        "level": {
+          "type": "string",
+          "enum": ["I", "II", "III", "IV", "V"],
+          "description": "The level at which this skill was unlocked."
+        },
+        "unlockedAt": {
+          "type": "string",
+          "format": "date",
+          "description": "Date when the skill was unlocked."
+        },
+        "unlockedIn": {
+          "type": "string",
+          "description": "Repository (owner/repo format) where the skill was detected."
+        },
+        "combinedFrom": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "pattern": "^[a-z][a-z0-9]*(-[a-z0-9]+)*$"
+          },
+          "description": "Parent skill IDs used in fusion (for composite/legendary skills)."
+        }
+      }
+    },
+    "pendingCombination": {
+      "type": "object",
+      "required": ["detectedSkills", "candidateResult", "levelFloor", "promptedAt"],
+      "additionalProperties": false,
+      "properties": {
+        "detectedSkills": {
+          "type": "array",
+          "items": {
+            "type": "string",
+            "pattern": "^[a-z][a-z0-9]*(-[a-z0-9]+)*$"
+          },
+          "minItems": 2,
+          "description": "Set of prerequisite skill IDs detected together."
+        },
+        "candidateResult": {
+          "type": "string",
+          "pattern": "^[a-z][a-z0-9]*(-[a-z0-9]+)*$",
+          "description": "The composite/legendary skill that would be unlocked."
+        },
+        "levelFloor": {
+          "type": "string",
+          "enum": ["I", "II", "III", "IV", "V"],
+          "description": "Minimum level required for the fusion."
+        },
+        "promptedAt": {
+          "type": "string",
+          "format": "date",
+          "description": "Date when the combination was first detected and prompted."
+        }
+      }
+    },
+    "treeStats": {
+      "type": "object",
+      "required": ["totalUnlocked", "highestRarity", "deepestLineage"],
+      "additionalProperties": false,
+      "properties": {
+        "totalUnlocked": {
+          "type": "integer",
+          "minimum": 0,
+          "description": "Total number of skills unlocked."
+        },
+        "highestRarity": {
+          "type": "string",
+          "enum": ["common", "uncommon", "rare", "epic", "legendary"],
+          "description": "Highest rarity tier among unlocked skills."
+        },
+        "deepestLineage": {
+          "type": "integer",
+          "minimum": 0,
+          "description": "Maximum DAG depth across all unlocked skills."
+        }
+      }
+    }
+  }
+}
